+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/AndroidAltice.iml" filepath="$PROJECT_DIR$/AndroidAltice.iml" />
       <module fileurl="file://$PROJECT_DIR$/MyApplication03.iml" filepath="$PROJECT_DIR$/MyApplication03.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
