+++ app/src/main/res/layout/activity_elegir_metodo.xml
@@ -20,7 +20,7 @@
         android:layout_height="match_parent"
         android:id="@+id/app03"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="example.ingva.com.myapplication03.Elegir_metodo"
+        tools:context="example.ingva.com.myapplication03.ElegirMetodoActivity"
         >
 
         <ImageButton
