+++ app/src/main/res/layout/activity_imc.xml
@@ -21,7 +21,7 @@
         android:layout_height="match_parent"
         android:id="@+id/app"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="example.ingva.com.myapplication03.imc"
+        tools:context="example.ingva.com.myapplication03.ImcActivity"
 
         >
