+++ app/src/main/res/layout/content_complexion.xml
@@ -29,7 +29,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="example.ingva.com.myapplication03.complexion"
+    tools:conteexample.ingva.com.myapplication03.ComplexionActivityActivityxion"
     >
