+++ app/src/main/res/layout/content_elegir_metodo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="example.ingva.com.myapplication03.Elegir_metodo"
+    tools:context="example.ingva.com.myapplication03.ElegirMetodoActivity"
     >
 
     <include
