+++ app/src/main/res/layout/fragment_enviar_comentario.xml
@@ -68,7 +68,7 @@
 
 
                 <Button
-                    android:id="@+id/button2"
+                    android:id="@+id/button_enviar_comentarios"
                     android:layout_width="wrap_content"
                     android:layout_height="0dp"
                     android:layout_gravity="center"
